Build: de-couple version and time-stamp
We want to handle the version and time-stamp (+git hash) differently.
If the version is updated, we want to re-build, whereas we want to
update the time-stamp only when at least one source is re-compiled.
(in other words, the time-stamp and git-hash should not trigger the
re-build.)

De-couple them in the build system.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>