Build: fix commit ID string when specifying out-of-tree building path
When the out-of-tree building path is specified by O=<path>, the Make
tries to get a commit ID of the repository on the path with 'git
describe' command, and apply the string indicating the ID to macro
GIT_HEAD.

If <path> is under other repository, ID string results in getting
from a commit ID of its repository. And if there is no repository on
<path>, ID string becomes an empty.

To apply own commit ID correctly, this patch adds changing the
current directory to the source directory in front of 'git describe'
command.

Signed-off-by: Kunihiko Hayashi <hayashi.kunihiko@socionext.com>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>