LD11: set CPLL to 1960MHz before CA53 clock gear up
Commit 6f717c1e4600 ("LD11: set CA53 frequency to 1100MHz") changed
the CA53 clock frequency up to 1000MHz (the commit subject is wrong!),
but the operation over 980MHz is not guaranteed.

Set the CPLL frequency to 1960MHz before setting CA53 clock to CPLL/2.
With this, the CPU will run at the maximum clock at the very beginning
of the boot sequence.

Reported-by: Dai Okamura <okamura.dai@socionext.com>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>