zlib: use malloc() instead of stack for workspace
Currently, the gunzip() allocates the zlib workspace (>9KB) in stack.
In the default configuration, 0x30010000-0x30014000 is used for the
.bss and stack, but this turned out to be a problem; the LD11 ROM uses
0x30010000-0x300120000 as the USB boot workspace.

Some temporary data are saved in this area, if it is corrupted, the
next loader (i.e. ARM Trusted Firmware) fails to load images from a
USB device by using the ROM API.  Currently, it is working, relying
on the fact the boot ROM has enabled the D-cache, and the .bss and
stack of this firmware will reside on cache lines.  This is not stable.

Every SoC has a reasonably large (>11KB) free memory in SRAM, but the
address varies SoC by SoC.  This firmware is designed to support all
SoCs by a single configuration, so it is impossible to specify such a
large memory arena by the CONFIG option.

One solution is to reduce the .bss and stack area to 2KB, then move it
to 0x3000f800-0x30010000, which we know is available for all SoCs since
it is the tail of the memory dedicated to this firmware.  Instead, the
concatenation of this firmware and the next image must fit in 62KB (it
was previously 64KB).

The gunzip() requires work memory more than 9KB, so it should be moved
to the heap, which is now available, and run-time initialized depending
on the SoC specification.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>