mbr: fix MBR overwrite issue for Windows 10 RS2 or later
Windows 10 RS2 or later forcibly overwrites the Disk ID field
(byte 440-443) in the MBR when it detects a disk with blank Disk ID.
This must not happen because the pre-computed signature will unmatch,
then the boot ROM will fail to verify the first stage loader.  The
disk ID field must be filled with something.

Signed-off-by: Dai Okamura <okamura.dai@socionext.com>
Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>