head: disable MMU at the very beginning of the entry
The boot ROM has enabled the MMU, but there is a lot of unknowledge
about memory attributes, address ranges, etc.

In fact, the MMU set up by the PXs3 boot ROM would cause a nasty
problem for the suspend-to-RAM.

Disable it at the very beginning of the entry code in order to start
our program at a clean state.  I did not see noticeable performance
regression.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>