mbr: move __packed marker after 'struct'
__attribute__((...)) for struct type can be placed either after
'struct' or after the closing brace.  The former syntax turned out
to be safer.

A problem recently root-caused in Linux kernel was, attribute
shorthand placed after the closing brace can result in a variable
name if it is not defined in time. [1]

Here, __packed is the same pattern; if __packed were not defined,
the compiler would interpret it as a variable name.  This case would
still trigger a build error, but a potential issue that would end up
with silent bad code generation should be avoided.

Linux commit 28128c61e08e ("kconfig.h: Include compiler types to
avoid missed struct attributes") chose to force the definition,
but I hesitate to include <compiler.h> globally.

I relocated the markers to the _less_ fragile position.

[1] https://lkml.org/lkml/2018/2/21/1403

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>