Build: do not redirect the first prerequisite for filechk
Currently, filechk unconditionally opens the first prerequisite and
redirects it as the stdin of a filechk_* rule.  Hence, every target
using $(call filechk,...) must list something as the first prerequisite
even if it is unneeded.

'< $<' is actually unneeded in all the callers.

Refer to Linux commit 43fee2b23895 ("kbuild: do not redirect the first
prerequisite for filechk").

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>