stdarg.h: import from Trusted Firmware-A v2.0
stdarg.h is the only header for which we rely on the compiler.
Given that some simple va_* defines are enough for us, it shouldn't
hurt to have a copy in this project.

This header has been imported from Trusted Firmware-A v2.0 [1].
Its license, BSD-3-Clause, is compatible with GPL v2, which this
project is licensed under.

[1] https://github.com/ARM-software/arm-trusted-firmware

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>