Build: add a dummy rule of the main source file for if_changed_dep
Unlike Linux, this project uses a simplified if_changed_dep because
there are only a few CONFIG options, and they are rarely changed. It
should not hurt to rebuild the whole tree when any of CONFIG options
is changed. That is why this project has not pulled in fixdep.

Instead of using fixdep, the dependency output from the preprocessor
is simply appended to the .*.cmd file. It mostly works well.

The -MP option adds a phony target for each of included headers.
These dummy rules work around errors when header files are removed.

Since -MP does not generate the dummy rule for the main source file,
a problem arises when the source file is changed from .c to .S or
vice versa. The straightforward solution is to add a dummy rule for
the main source file.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>