entry: use a pointer to struct board_data as an entry symbol
The entry point (-e option of the linker) does not need to be a
function. Use the pointer to struct board_data as an entry symbol.
It still gets along with the dead code elimination.

This is somewhat hacky, but the motivation of this change is to get
access to board_data->soc_data before setting the stack pointer.
This will allow the stack to use larger memory.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>