stack: reuse heap memory for stack
Every SoC has somewhat large memory arena in SRAM.

 LD11: 0x30012000-0x30014c00
 LD20: 0x30016000-0x3001c000
 PXs3: 0x30016000-0x30020000

The location varies SoC to SoC, so it is currently only used for heap.
The stack pointer is set to the fixed address, 0x30010000, irrespective
of the SoC variant. This reduces the precious memory footprint.
Remember we can use only 64KB for uniphier-bl + bl2.bin.gz

Here, a new gimmick. First, set up a temporary stack for calling
get_stack(). The first 512 byte is just a copy of the MBR of the boot
device, so it can be used as a stack memory.

get_stack() returns the end of the large work memory, which sp is
set to for running the rest of the code.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>