linker_script: add image size check
There is no guarantee if the SRAM can be used beyond the 64KB offset.

It surely can for LD20 and PXs3 since they load 80KB for the first
stage boot loader. LD11 reserves 0x3001000-0x30012000 for the BIOS
APIs for USB loading, so the .bss section should not overrun the 64KB
limit.

Let's add ASSERT() in the linker script.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>