treewide: convert to SPDX License List v3 tags
GPL-2.0 and GPL-2.0+ are deprecated:

  https://spdx.org/licenses/GPL-2.0.html
  https://spdx.org/licenses/GPL-2.0+.html

While they are still valid and are widely used in Linux, the recent
SPDX conversion storm in Linux is based on SPDX License List version 3.

  https://lkml.org/lkml/2019/5/22/237

Linus Torvalds called whoever cared about this "a hopeless wanker"
(https://lkml.org/lkml/2018/8/22/496), but the background of this
change is well-explained in this article:

  https://www.gnu.org/licenses/identify-licenses-clearly.html

Anyway, this project is small enough. I just converted all tags by sed
before consuming time for this argument.

  GPL-2.0  -> GPL-2.0-only
  GPL-2.0+ -> GPL-2.0-or-later

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>