Support position independent executable
This loader was designed relying on the assumption that the Boot
ROM loads the first stage loader to the same address, 0x30000000.

For the planned new SoC, the load address will be 0x04000000.

The build system is able to generate all SoC/board binaries without
per-target configuration, and I want to keep this design.

To support the new SoC and existing SoCs in the single configuration,
this commit supports PIE.

When CONFIG_PIE is enabled, the position independent executables are
produces. The binaries can be loaded anywhere irrespective of the
link address.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>