Split init.h out into small headers
For the planned new SoC, the base registers of IP blocks will be
completely changed.

The base addresses will be embedded in struct soc_data. This means
most of source files are required to include <init.h> to get access
to the hardware.

I do not want to make this header "a collection of all sort of
declarations included by everyone". Split it out into small chunks.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>