Build: remove .bss section explicitly for llvm-objcopy
I noticed a breakage when building with the latest LLVM tools
due to the behavior change of llvm-objcopy.

The first bad commit in the LLVM is:

|commit 2b68379b89263178917e3bd37eaf9154078c2965 (HEAD)
|Author: Peter Collingbourne <peter@pcc.me.uk>
|Date:   Fri May 24 00:21:46 2019 +0000
|
|    llvm-objcopy: Change sectionWithinSegment() to use virtual addresses instead of file offsets for SHT_NOBITS sections.
|
|    Without this, sectionWithinSegment() will return the wrong answer for bss
|    sections. This doesn't seem to matter now (for non-broken ELF files), but
|    it will matter with a change that I'm working on.
|
|    Differential Revision: https://reviews.llvm.org/D58426
|
|    git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@361578 91177308-0d34-0410-b5e6-96231b3b80d8

Prior to that commit, the size of the final binaries was 20KB or so,
now it is 65280 byte.

Looks like the latest llvm-objcopy does not rip off the .bss section.

Add '-R .bss' to work around the issue.

I filed a bug report:
https://bugs.llvm.org/show_bug.cgi?id=42277

Once this problem is solved in the upstream, I will revert this commit.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>